add weekly directories
