+++ Week 1/Vehicle.py
@@ -5,14 +5,15 @@ class Car:
     #     self.model = None
     #     self.year = None
 
-    def __init__(self, color=None, brand=None, model= None, year = 0) -> None:
+    def __init__(self, color=None, brand=None, model=None, year=0) -> None:
         self.color = color
         self.brand = brand
         self.model = model
         self.year = year
 
+
 class Truck:
-    def __init__(self, color=None, brand=None, model= None, year = 0) -> None:
+    def __init__(self, color=None, brand=None, model=None, year=0) -> None:
         self.color = color
         self.brand = brand
         self.model = model
